refactor: 영상 분석 리스트 컴포넌트를 공통 컴포넌트로 분리
